[cpu-daemon & kernel] add option for pss or meminfo usability
[cpu-daemon description]
PSS is commonly not imported on user variant build.
And it takes a lot of time to calculate pss of each processes.
Furthermore, it is found that CPU usage and its frequency go to max while
calculating PSS.
This will be default off and if you want to enable using PSS calculation,
echo 1 > /sys/devices/system/cpu/cpu0/cpufreq/use_pss

[kernel description]
CPU frequency change in ondemand governor will be monitored in the kernel logs.

The frequency goes to max like
[DBG UP to MAX] avg_freq : 200000, load_freq : 19800000, load : 99, cur_freq : 200000, up_t_freq: 19000000, up_t : 95

The frequency goes gradually down like
[DBG DOWN] avg_freq : 1704000, load_freq : 25560000,  load : 15, cur_freq : 1704000, down_t_freq: 156768000, down_t : 92, next_freq : 277826
[DBG DOWN] avg_freq : 300000, load_freq : 2400000,  load : 8, cur_freq : 300000, down_t_freq: 27600000, down_t : 92, next_freq : 200000